requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
requests
httpx
openai
replicate
yt-dlp
supabase
python-jose[cryptography]
